refactor: extract StaffRecordsTab from admin index
Move the staff records tab render function into a dedicated
component that receives all needed data and callbacks as props.
